Convert lastSectorTimestamp to millis timestamp and do tests

+++ include/gps/gps.h
@@ -26,7 +26,7 @@ int getLapCount();
 
 float getLastLapTime();
 
-float getLastSectorTime();
+unsigned long long getLastSectorTime();
 
 int getSector();
